(CFACT-204) Add request object to HTTP exceptions.
The exception thrown by the HTTP client when there is a failure to
connect or process a response from the server should have a copy of the
request that caused the failure.  That will allow code that makes many
different HTTP requests be able to determine which request failed.

The request and response objects had previously explicitly implemented
move semantics because the copy semantics were deleted.  Because we now
need the copy semantics, we can use the default copy and move semantics
for these types.