(CFACT-204) Fix EC2 instance detection for PV instances.
The EC2 fact resolver did a check for "amazon" in the BIOS vendor in an
attempt to prevent an unnecessary network operation when running
under kvm/xen* hypervisors.  However, the BIOS information is only
available to Hardware Virtual Machines (HVM).  For paravirtual (PV)
instances, the BIOS information is unavailable, causing the EC2 resolver
to skip resolving EC2 facts.

As there is no reliable way to detect if we're running under an EC2
instance without querying the metadata endpoint, the fix is to remove
the BIOS vendor check and gracefully handle failing to query the first
endpoint URL.