(CFACT-208) Accept RPC_E_CHANGED_MODE from COM
When run in Puppet, cfacter would fail to initialize COM because it
returned RPC_E_CHANGED_MODE. We don't care what the threading model is, so
this should be considered a success.

Changes COM initialization to accept RPC_E_CHANGED_MODE as having COM
initialized and continue with WMI calls. Doesn't call uninitialize,
because COM doesn't see it as a successful initialization.