(CFACT-206) Fix crash in JSON output for ruby hash with non-string key.
If a custom fact resolution returned a hash with a non-string key
(typically a symbol), then Ruby throws an expcetion from
rb_string_value_ptr because the given VALUE is not a string.

The fix is to properly convert the key to a string like we're already
doing for the other formats.