(CFACT-180) Strip nulls when reading registry
Reading strings from the registry is returning a null terminator.
Most of the time this wasn't obvious (the output would include an
extra null, but appear visually identical). If the string should
have been empty however, it would include an extra null character,
leading the domain fact lookup to accept it as a non-empty string.

Strips trailing null character, and calculates string length
correctly so that domain fact lookup works.