(CFACT-198) Remove cfacter gem and FFI dependency.
Previously we had a gem/ directory that defined the cfacter gem.
This was no longer being used as the Ruby API is actually implemented
entirely in libfacter (i.e. in C++).  We kept it around to install
cfacter.rb, but that was being done independently of the gem.

This commit removes the gem entirely and moves the specs over to the
lib/ directory.  Additionally, the FFI dependency has been removed in
favor of having Ruby directly require libfacter as an extension.