(CFACT-219) Limit to prior Facter external facts
On Windows, Facter checks PATHEXT, but then always checks whether the
files pass File.executable?. Ruby hard-codes only .com, .cmd, .exe, and
.bat to be executable on Windows, so PATHEXT is essentially ignored.

There is a corner cases where this will give different behavior: if
PATHEXT is set but doesn't include one of the 4 extensions mentioned
above, prior Facter implementations would not try to execute files with
the excluded extension. In the current implementation, we still execute
them because they are still executable in the Windows environment.