(FACT-822) Prioritize AIO private bin over PATH for virt-what
Without this change, facter uses PATH to find virt-what. The aio
puppet-agent will include its own version of virt-what in a location
that won't be in PATH, that's intended to be used by facter.

Override the lookup for virt-what to first try the aio path containing
virt-what, and fall back to PATH if not found.