(CFACT-217) Exit with a failure code when an error has been logged
Previously, cfacter would only exit with a failure code if an
exception managed to bubble all the way up to main. This meant that it
was impossible to tell without parsing the log output whether an error
had actually been encountered.

This adds functionality to the logger to track when an error or fatal
message is logged, and updates the cfacter main to use this
functionality to return a failure code when such a message has been
loged.