(CFACT-126) Rename 'productuuid' top level flat fact to 'uuid'
In native Facter, the former 'uuid' top-level fact was renamed to 'productuuid.'
To maintain backwards compatibility, this commit reverts that change so
that 'uuid' is once again the top-level fact name.

Note that top-level flat facts are now considered legacy facts, since the
default output of native facter is fully structured. This change does
not affect the 'dmi' structured fact, or the 'dmi.product.uuid' fact
within it. It can only be accessed directly by invoking cfacter with
'uuid' as an argument.