(CFACT-230) Fix architecture fact resolution on certain Linux distros.
On Debian, Gentoo, kFreeBSD, and Gentoo, the architecture fact should
resolve to "amd64" for "x86_64".  On Gentoo, it should resolve to "x86"
for "i386".  However, the code was attempting to resolve the
hardwaremodel fact to these values instead.  Additionally, the check for
the distro name occurs before it was being populated.

The fix is to move the check to after the distro name is being resolved
and change the assignment to the architecture fact.