(CFACT-231) Resolve processors.isa and hardwareisa in Windows
Prior to this commit, the processors.isa and hardwareisa facts were
not being populated in Windows.

This commit adds the necessary logic to query WMI for "Architecture"
and transform the returned numerical ID into a string representing the
processor type.

Details on the translation from ID to processor type can be found here:
https://msdn.microsoft.com/en-us/library/aa394373%28v=vs.85%29.aspx