(CFACT-154) Merge selinux resolver with OS resolver.
Rather than creating a base resolver for this platform-specific
resolver, merge the implementation in with the OS resolver as this is
information about the operating system.

This adds more structured data to the "os" fact.  There is now a
"selinux" child that is present on Linux systems and contains data
relating to selinux.

This also corrects the type of the "selinux" and "selinux_enforced"
facts, which should be boolean according to Facter 2.x schema.