puppetversion: massive speed boost
Running facter with --timing shows that 'puppetversion' is by far the slowest
fact. This is not true when executed by Puppet, but gives users the wrong
impression that Puppet is sooo slow. This can easily be fixed by not half-way
bootstrapping Puppet unless --puppet is used.

This fix has has an influence on execution speed for standalone Facter, you'll
not notice a difference when calling Facter while Puppet is already running.