(maint) Update beaker pin and allow env var override
This adds support for BEAKER_VERSION as was added in puppet at commit:

https://github.com/puppetlabs/puppet/commit/072891f097fd4fe7e35d93965c9b3c06d78bf245

However, the beaker pin is to master, since it has updated AIO paths.

This change also updates the multi-json gem version - this was pinned to 1.7.7
over a year ago to avoid a bug since fixed, so we can return to normal
pessimistic versioning. This is necessary because newer beakers pick up a
requirement for a newer version of multi-json via a transitive dependency.