(FACT-825) Increase prtdiag timeout for Solaris virtual fact
This commit bumps up the default timeout for the virtual fact in
Solaris to 20 seconds. `prtdiag` on large Solaris systems can often
take upwards of 10 seconds, so 20 seconds is a reasonably safe timeout.

This fix was previously merged (d82f695a0) and then lost in the
Facter-2 / master branch merge.