(CFACT-235) Pull in leatherman options earlier in CMakeLists.txt
Previously we got some CMake options as part of the leatherman cflags
include. This was too late in the CMake code to affect find_package
calls that we expected to respect these options. Specifically,
BOOST_STATIC was doing nothing of note.

We now bring in the options fairly early (after we set up our own
options, but before we find any packages)