(CFACT-132) Handle link-local addresses consistently
Facter completely hid link-local addresses (127.0.0.1, fe80::). This was
likely as a mechanism for not using them in any decision-making.

In CFacter, we chose to show the addresses, but still avoid using them as
a mechanism for selecting the primary interface. This should have no
change in behavior other than printing out additional facts showing the
link-local addresses.

Windows CFacter wasn't originally implemented to conform to that change;
it reproduced the original Facter behavior. Change the Windows
implementation to mirror other platforms and show link-local addresses,
but ignore them for selecting the primary interface.