(CFACT-253) Implement timeout option for POSIX command execution.
This implements a timeout option for POSIX command execution by using
the POSIX timer API to mark a "timed out" flag and change the read
call to be a select call with a timeout to check the flag.

When a command times out, a facter::execution::timeout_exception will be
thrown.

Also fixing reaping child processes when an exception is thrown.  We
previously failed to reap the child if a read error occurs.  We now kill
the child and reap it when an exception (including a timeout) occurs.