(maint) Reduce build procs to avoid out-of-memory
Debian build VMs are occasionally running out of memory with the message
`virtual memory exhausted: Cannot allocate memory`. This is likely
because compiling C++ is a memory hog, and machines don't have enough
memory allocated to cover a build for each core.

Hard-code the number of parallel procs to 2. Most of our debian builders
use 4GB of memory, and more than 2 processes are likely to exhaust that
at some point.