(CFACT-264) Execution timeout on Windows kills process tree
Command timeout for Windows systems does not properly kill the entire
child process tree, only the process directly spawned by libfacter.

Modify the unit test to demonstrate this problem by calling ruby from a
command shell, so we have a two-level process tree. With that modification
and without the rest of this commit, ruby.exe is not terminated and takes
the unit tests are hung for 60 seconds.

Create a job object and associate the spawned process with it, then use
TerminateJobObject to kill the entire process tree.