(CFACT-268) Don't resolve custom facts when --no-custom-facts is used
This commit fixes a regression around the --no-custom-facts
command-line option. Currently, the option is not being honored,
and so custom facts will be resolved even if the option is used.
This commit ensures that custom facts will not be resolved if either
this option is passed or if Ruby is not present.