(maint) Fail hard if building native facter throws errors
We are currently running into some errors where building native facter
will throw errors, but the script will continue to run. This isn't good,
especially because it means we ship an incomplete archive. This commit
is an attempt at having the native facter build script fail in a
consistent way when things go wrong.