(FACT-855) Test facter output against the schema
Tests facter output against the schema to confirm that the reported
facts are in the schema and match the specified type.

Explicitly requests known hidden facts to confirm they can be retrieved
as well, and validates them against the schema.

Also fix an incorrect comment in unit test schema validation.