(FACT-865) Split stdout and stderr handling in execution.
This implements a new pipe for reading stderr in addition to stdout from
child processes.  Previously, stderr was optionally redirected to
stdout or to the null device if not merging the two streams.  This change
allows child process execution to treat stdout and stderr as two
separate streams, enabling the ability to process stdout while logging
stderr messages (e.g. external executable facts).

The execute functions now return a tuple of both the stdout and stderr
contents.  The each_line functions now take an additional function
object to call when a line of stderr is read.