(CFACT-258) Fix Facter::Core::Execution#execute not raising by default.
If an options hash was passed to Facter::Core::Execution#execute and the
options hash did not contain an :on_fail key, we were not properly
defaulting the value to :raise to raise an error.

This fix calls rb_hash_lookup2 to supply :raise if the lookup fails,
resulting in raising an exception if the :on_fail key is not present in
the options hash.