(docs) Update submodule instructions to work after clone
Users often clone the repo before reading build instructions; or they
may clone, then checkout a branch, and want to ensure the submodules are
updated for that branch.

Change `git clone --recursive` to `git submodule update --init` to
ensure submodules are updated for the current branch, and provide
instructions that work after cloning.