(maint) Ensure AIO pre-suite uses PC1 repository
Prior to this we were installing repositories for products/devel or main
(for rpms and debs respectively).

Now AIO packages will be in a repository called PC1. This adds checks to
the repository set up steps so that PC1 will be tried first. Once
packaging is fully switched over will need to remove references to older
repositories.