(FACT-900) Fix external facts path for non-Administrator
The external facts path for non-Administrator users was incorrect, it used
COMMON_APPDATA when the puppet specification states it should use
~/.puppetlabs/opt/facter/facts.d. One reason for this location is that
facts in COMMON_APPDATA may not be executable by non-Administrators.

Determine whether Facter is running with Administrator privileges; if not,
use the location in the HOME directory instead of COMMON_APPDATA.

This commit implements a number of utilities from Puppet and Ruby that
we'll want to move to Leatherman, so we can share the implementation.