(FACT-904) Fix external fact load ordering.
Previously, libfacter was searching for external fact files to load and
mapping those files to the resolvers that can handle them.  This mapping
was done with a std::map, which resulted in a lexographical ordering
based on the path of the file when the loading was performed.  Users
expect facter to load external facts based on the order the external
directories were specified.

Therefore, this commit removes the unnecessary mapping and simply loads
an external fact file when it is found.