(maint) Check stderr for output in run_external_facts_once.rb.
When the master branch was updated to native facter, we temporarily
replaced the check for external fact script output on stderr with a
redirection to a file.  This was because native facter ignored output
on stderr for external facts.  A recent change now outputs the error as a
warning on stderr, so this commit reverts the previous workaround.