(FACT-947) Use GetUserProfileDirectory
Previously, we queried environment variables to find the User's profile
directory. This followed logic used in Puppet to do the same. That was
more error-prone and required querying multiple environment variables.

Use GetUserProfileDirectory to get the path instead, which should work on
all platforms. In particular this will be better supported in future
Windows versions.