(FACT-896) Fix link errors on Solaris.
libfacter_test currently doesn't link on Solaris.  This was caused by
the changes to make symbols in libfacter private (the test executable
now links in the object files rather than libfacter).

As a result, the test executable can't resolve inet_ntop, which is
defined in libnsl.  The reason this "works" for the facter executable is
because it links against libfacter.  When libfacter loads and resolves
the inet_ntop import, it just happens to find one because libnsl has
already been loaded.  This won't work for linking the test executable
because the symbol needs to be defined at link time.

The fix for this is to properly link in libnsl.  Also removing kstat and
socket from the global libraries to link and instead adding them to the
platform libraries in libfacter/libfacter_test, where they belong.