(maint) Remove unnecessary check
The check for ERROR_NO_SUCH_PRIVILEGE came from
https://github.com/puppetlabs/puppet/blob/master/lib/puppet/util/windows/process.rb#L200,
and reviewing the MSDN documentation of the APIs further I don't see any
place that error should be generated.

It would likely be on the call to GetTokenInformation (where TokenElevation
may not exist on 2003). This code is never called on 2003 because it's guarded
by supports_elevated_security, so having the debug message shouldn't hurt.