(FACT-697) properly skip fact if NetworkManager is not running.
On my CentOS 7 nmcli printed an error to STDERR if NetworkManager
is not running, however this was redirected to /dev/null
This had effect that the confine didn't work and facter tried to
exectue `nmcli d`, which then printed the same error to STDERR,
having the mesage "Error: NetworkManager is not running." on each
facter/puppet run.

We now properly check if something at all was returned and
otherwise assume that nmcli won't work anyways.