(FACT-955) Fix hardwaremodel/architecture fact for WOW64.
On 64-bit Windows, facter was reporting "i686" for the hardwaremodel and
architecture facts when running under WOW64 (i.e. facter itself was
32-bit).  There was an incorrect check in Facter for WOW64 that was
reporting the architecture of Facter and not the operating system.

This fix removes that check.  Additionally, removing architectures that
are no longer supported (Windows only supports x86, x64, ia64, and ARM).