(FACT-981) Fix Ruby API confine to allow for rspec mocking.
In Facter 2.x, the confine logic would call back into the Facter API to
retreive the fact to test the confine against.  This allowed module
authors to test their custom facts by mocking Facter#fact or
Facter::Util::Fact#value.

In Facter 3.x, the confine logic is implemented entirely in native code.
Because it does not call back into the Ruby API to obtain the fact's
value, the mocked values aren't respected and the confines keep to the
original fact values.

This change alters the confines to call back into the Ruby API to
retrieve the fact.  This enables mocking of the API for module tests.