(FACT-981) Stop freezing Facter modules.
Originally native facter freezed modules such as
Facter::Util::Resolution so that it could find places where Facter was
being modified. However, this breaks being able to mock parts of the
Facter API for Puppet module testing.

This commit removes the freeze calls so that the API can be stubbed.