(FACT-96) Deprecate `facter --puppet`
The `--puppet` option in facter introduces a cyclic dependency between
the two projects, and adds complexity, as facter needs to know how to
initialize puppet.

This commit causes `facter [-h|--help]` to mention that the `--puppet`
option is deprecated. It does not emit a deprecation warning when
running `facter --puppet` because that command is run often on every
agent. It also adds tests that the warning is printed.

The recommended way to retrieve puppet-specific facts is to run `puppet
facts find`, which includes `puppetversion`, and pluginsync'ed custom
and external facts.