(FACT-96) Deprecate `facter --puppet`
The `--puppet` option in facter introduces a cyclic dependency between
the two projects, and adds complexity as facter needs to know how to
initialize puppet.

This commit adds a deprecation warning when executing `facter --puppet`
and in viewing facter help. The alternative is to run `puppet facts
find`. It already handles puppet-specific facts, e.g. puppetversion,
custom facts in `$vardir/lib/facter`, and external facts in
`$vardir/facts.d`.

Note the deprecation warning is not triggered if code loads facter as a
library and calls `load_puppet` directly. Trying to account for that
case seems like more trouble than it's worth.