(FACT-628) Move LDOM virtual implementation to virtual.rb
Previously, the implementation of the `virtual` fact that evaluated
LDOM info was stored in `ldom.rb`. This would cause the value of
`virtual` to be different depending on whether or not ldom.rb was
loaded.

This moves the implementation of the virtual fact to virtual.rb, and
weights it appropriately within the set of implementations there.