(FACT-892) Change "POSIX" load average resolver to "glib".
The base load average resolver implementation is currently building as
part of the POSIX resolvers, but it calls getloadavg, which is not a
POSIX function.  It is actually a glib function.

This breaks the build on systems where "getloadavg" isn't part of the
standard library implementation, such as AIX.

This moves the resolver into a "glib" namespace and adds the resolver to
instances where a conforming glib is present (OSX, BSD, Linux, and
Solaris).