(FACT-976) Remove parse_release step
Prior to this commit, parse_release was used to parse the release string, and
overridden as needed. For Windows and Solaris, using the parse_release format
actually made the work more complicated. On Linux, some distros use different
parsing, and the method of determining distro was reworked for FACT-976 in a
way that makes it complicated to pass to the parse_release method.

Instead of providing a parse_release method, parse the release major/minor
parts when the release string is determined and store them to the same data
container.