(FACT-965) Ensure fact resolution occurs under JNI lock for JRuby.
Currently, fact resolution occurs upon the first call to
`com.puppetlabs.Facter.lookup`. Because native libraries are shared between
Java threads and fact collections are not thread-safe, this is inherently unsafe.

This moves the resolution of the facts into loading of the library,
which is protected by a JVM mutex.  Puppet Server will now load the facter
JNI shim class at startup, causing all fact resolution to occur only
once at startup.  Because the JNI shim only supports reading a fact's
value, subsequent calls to `com.puppetlabs.Facter.lookup` do not need to
be protected as the collection is considered to be immutable after
loading of the native library.