(FACT-967) Go back to oslevel-style OS version facts on AIX
Feedback from test pilots was that the facter-2 output was more in
line with what they expected than in the first pass at facter-3. This
commit brings facter 3 in line with that old style output.

Rather than shelling out to the very slow oslevel, we instead use
oslevel's cache file ourselves when it's available. This takes about
0.5s on our test AIX box, vs 2s to run oslevel. If the cache doesn't
exist we still need to call out to oslevel to build it, but we
shouldn't end up taking that performance hit very often.