(FACT-1044) Refactor kernel resolver version parsing
This commit refactors version parsing in the kernel
resolver by creating a new parsing utility function
to be used to return major and minor versions via
`collect data.` In addition, this commit adds a new
Linux kernel resolver which uses a new regex to parse
kernel version. This became necessary when `uname` output
changed slightly in Debian 8, no longer including the
patch level for the kernel.