(FACT-1073) Add major/minor OS release facts for distros that lack them
Previously, if a given OS did not have a proper OS release, we would
use the kernel release. This release would never be split into
major/minor versions.

This now splits that kernel release into major/minor versions, as
Facter 2 did. For platforms that have their own os release
implementation, this will continue to behave as it did before.

Updating the test for this is a simple matter of removing major/minor
from the fixture data so that we can see it has been calculated by the
resolver.