(FACT-1084) Catch failures to convert address
Without this change, Facter calls WSAAddressToStringW with an invalid
SOCKET_ADDRESS and throws an uncaught exception when the call fails. This
leads Facter to exit ungracefully.

Ensure calls to WSAAddressToStringW catch any exceptions thrown as a
result of errors, so we simply fail to resolve the address.

Also extend allowed IP address length on Windows to account for null
terminator, and skip resolving dhcp if not enabled.