(FACT-1083) Replace fork/execv with vfork/execve.
To assist running facter from puppetserver where it is consuming a large
amount of memory in a constrained environment, we should use vfork over
fork so that the parent process' page tables aren't copied into the
child.  This allows facter to spawn additional processes from
puppetserver even when there isn't enough memory to commit to a forked
puppetserver process.

The downside of vfork is that it is a very unsafe system call.  It
creates a child process like fork does, but instead of copying the
parent process' page table, the child process shares the parent process'
address space entirely.  That means that the child can very easily
corrupt the parent.

Therefore, it is with great care that we control exactly what code the
child process runs.  It should not modify any program state, allocate
memory, throw exceptions, or do anything other than duping and closing
file descriptors before calling execve.  Also note that unlike fork,
vfork implementations suspend the parent until the child calls an exec
function or _exit.  This is so that the child can call functions and
append to the current frame's stack space (it must modify the stack
above the current stack pointer, though).

Note: we no longer need the hacky ruby::api::cleanup flag because the
child process now calls _exit when it fails to execve, which prevents
global destructors from running in the child process; therefore
preventing ruby from attempting to shutdown from the child process.