(FACT-1085) Fix Facter execution API redirecting stderr to stdout.
The Facter execution API is currently redirecting stderr to stdout.
This means that commands that fail will return the error output in
addition to any other output in the return value.  This is a regression
from Facter 2.x.

The fix is to redirect stderr to null.  Users can still see the error
output when passing --debug to facter.  Additionally, users can manually
redirect stderr output to stdout in their commands using shell
redirection syntax (e.g. `2>&1`).