(FACT-1085) Fix Facter command execution for nonexistent commands.
When Facter::Core::Execution::exec is called with a nonexistent command,
Facter 2.x returns nil instead of empty string.

When Facter::Core::Execution::execute is called with a nonexistent
command, Facter 2.x raises a command not found error by default.

This commit ensures that Facter 3.x conforms to that behavior. The
problem was that facter::execution::execute was expanding the command
given to it, which in this case was the shell, so it succeeded. The fix
is to ensure the command expanded before attempting to execute it and
replicate the logic in the previous ruby implementation.