(FACT-1090) Query Ruby `$LOAD_PATH` when used
Facter 3.0 included a regression where running Facter in some Puppet
commands would not include the correct Ruby `$LOAD_PATH`, resulting in
missing facts. This happens because Facter caches the `$LOAD_PATH` when
initialized, and only resets it when `Facter.reset` is called.

Fix the regression by updating Facter to query `$LOAD_PATH` when facts
are loaded instead of caching on initialization.