(maint) Relax validation of uptime related facts
On OSX, the system profiler returns uptime as either '5 minutes' or
'1:23', where the latter is used if the system has been up for at least
an hour. The verify_facts acceptance test does a YAML.load of facter's
output, automatically converting '1:23' to 4980 (of type Fixnum), which
doesn't match the schema.

This commit relaxes the validation to allow uptime related facts to be
strings or fixnums. On OSX, this can happen with the
system_profiler/uptime fact and the legacy sp_uptime fact, so we regex
on the fact name.

Also outputs the value's type when we fail.

[skip ci]