(FACT-1111) Bring back `facter -p`
The beast rises from beyond the grave to once again terrorize the
local villagers.

This brings back our circular dependency on Puppet. When the
`--puppet` or `-p` arguments are passed to Facter, it will attempt to
load Puppet's ruby code in order to find additional facts that may
have been pluginsunc to the agent.