(FACT-1115) Fix execution of non-zero command exit treated as failure.
Facter 2.x returns output from `Facter::Core::Execution::exec/execute`
even if the command failed (non-zero exit code).  It returned the value
of the `on_fail` option only if the command couldn't be found.

Facter 3 was returning the `on_fail` option if the command returned a
non-zero exit code or couldn't be found.  This commit restores the 2.x
behavior.