(FACT-1123) Retry locale setup with LANG=C when we fail
Previously, if we failed to setup a locale (for any reason) we would
fail hard. With this PR, we now clear the locale-related environment
and try again, printing a warning message to the user.

If this second attempt fails, we'll still fail hard. We have no
logging if our locale setup has failed, so there's not a lot we can
do.

This also fixes a bug in the ruby module where it was initializing
logging incorrectly.