(FACT-1125) Include facter bin in PATH on windows
Facter was not loading on Windows because the path to the libfacter.so
is not present and Facter could not load itself.  This was affecting MCO
on windows in particular when trying to run the puppet agent plugin
which requires Puppet, which requires Facter.  The work around here is
to prepend the windows PATH environment variable with the directory
libfacter.so is located in.  We're looking it up relative to the
facter.rb file being executed, and the expectation is that the directory
structure on Windows should remain stable.