(FACT-1133) Fix unnecessary object allocation from Facter API.
When resolving a value for a built-in fact, the `_resolved` flag never
gets set, so the cached Ruby object for the built-in fact is thrown
away and a new object is allocated.  This is unnecessary and just
creates garbage.

The fix is to always set the flag to true when we're done resolving the
fact.