(maint) Confirm doxygen issues no warnings
Changes the doxygen test to use quiet output and scrape for any
warnings, as WARN_LOGFILE doesn't appear to work.

Also fix a few doxygen warnings. In particular, make Extensibility.md
part of the docs so we can use a relative reference.

Update Doxygen to 1.8.10; no strict requirement for this, only keeping
it up-to-date.