(FACT-1140) Fix decoding DMI data containing non-printable ASCII.
On Linux, we directly read DMI from data that is provided from the kernel.
On some hardware the bytes read are not printable ASCII, resulting in
invalid UTF-8 sequences when interpreted as UTF-8.

The dmidecode utility converts any non-printable character to a '.'
character.  This fix mimics the behavior of dmidecode so that invalid DMI
data is not converted to invalid UTF-8.