(FACT-1123) Signal to library consumers when locale setup fails a second time.
Previously, if we still failed locale setup after cleaning the
environment we would simply propagate that exception out to the
world. Now we raise a special LocaleError exception that consumers can
use to recognize that logging setup was unsuccessful and take
appropriate action.

In main(), this action is to log via stderr and exit with an error
code. In the ruby bindings, we will raise a LoadError and forward the
exception message to the interpreter.