(FACT-1152) Use /sys/block in the linux FS resolver
Previously /sys/class/block was used in the linux FS resolver. This is
not available on older linuxes such as el4/el5/sles10, which meant that
the size on partitions there would be returned as 0. /sys/block exists
in both older linux and newer linux, so this commit updates the linux
size function to read /sys/block instead of /sys/class/block.
The /sys/block varation requires deriving the device name for the
partition, which requires boost::regex.