(FACT-1147, FACT-1150) Fix network interfaces with multiple addresses.
Network interfaces with multiple assigned addresses were not being
handled properly.  Facter was considering the last address enumerated to
be the address of the interface.  This may not be correct as the order
may have a non-link-local address first and a link-local address later.
This would result in facter favoring the link-local address which is not
desired (FACT-1150).

Additionally, if the interface had only link-local addresses, the last
address was being selected rather than the first (FACT-1147).

The fix is for the resolvers to return an array of bound addresses for
each network interface. The base resolver now selects what it considers
to be the address of the interface by finding the first address that is
not link-local. If all addresses are link-local, then the first bound
address is used.

Additionally, if network interfaces have more than one bound address,
a new `bindings` (or `bindings6` for IPv6) element will appear on the
interface containing an array of structures containing the address,
network, and netmask for each binding.  This will allow users to see all
addresses on the interface.

This fix also includes returning a "primary" element in the "networking"
fact which is set to the name of the primary interface.  Users can use
this to quickly determine the interface that facter considers to be the
interface used for the top-level networking facts.